fix compiler warnings (#800)
* fix compiler warnings

* add free

+++ contrib/win32/win32compat/fileio.c
@@ -702,7 +702,7 @@ fileio_write_wrapper(struct w32_io* pio, const void* buf, size_t bytes_to_copy)
 	int bytes_copied = -1;
 	size_t chunk_size = 0;
 
-	for (int i = 0; i < bytes_to_copy; i += WRITE_BUFFER_SIZE, chunk_count++) {
+	for (size_t i = 0; i < bytes_to_copy; i += WRITE_BUFFER_SIZE, chunk_count++) {
 		chunk_buf = (BYTE*)buf + chunk_count * WRITE_BUFFER_SIZE;
 		chunk_size = ((bytes_to_copy - i) >= WRITE_BUFFER_SIZE) ? WRITE_BUFFER_SIZE : (bytes_to_copy - i);
 		bytes_written = fileio_write(pio, chunk_buf, chunk_size);
